# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Healthcare Automation API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the current backend implementation
    (backend/src/routes, controllers, services, validators, middleware).
    
    ## Authentication
    Most endpoints require header:
    `Authorization: Bearer <JWT>`
    
    Obtain JWT via POST /api/auth/login.
    Token claims: userId (number), role (string), email (string).
    Default expiry: 1h. With rememberMe: 30d.
    Secret: process.env.JWT_SECRET or fallback "local-development-secret".
    
    ## Roles (UserRole enum)
    ADMIN, DOCTOR, PHARMACIST, PATIENT, SUPPORT, VIEWER
    
    SUPPORT and VIEWER exist in the schema and can be assigned via admin user APIs.
    They are authorized for read-only audit access via GET /api/audit-events
    (along with ADMIN). Other business routes do not authorize SUPPORT/VIEWER.
    
    ## Base URL
    http://localhost:4000 (PORT hardcoded in backend/src/server.ts)
servers:
  - url: "http://localhost:4000"
    description: "Local backend (PORT=4000 in server.ts)"
...
paths:
...
  /api/medications:
    get:
      tags: [Medications]
      summary: List medications
      description: "Roles: ADMIN, PHARMACIST. Filters: status, stockStatus, q."
      operationId: listMedications
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - name: stockStatus
          in: query
          schema:
            type: string
            enum: [IN_STOCK, LOW_STOCK, OUT_OF_STOCK]
        - name: q
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Medications]
      summary: Create medication
      description: "Roles: ADMIN, PHARMACIST."
      operationId: createMedication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, unit]
              properties:
                sku: { type: string }
                name: { type: string }
                unit: { type: string }
                quantityOnHand: { type: integer, minimum: 0 }
                reorderLevel: { type: integer, minimum: 0 }
                reorderQuantity: { type: integer, minimum: 0 }
                status: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        201:
          description: Created
        400:
          $ref: "#/components/responses/ValidationError"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        409:
          description: Duplicate SKU
  /api/medications/{id}:
    get:
      tags: [Medications]
      summary: Get medication
      operationId: getMedicationById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Success
        404:
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Medications]
      summary: Update medication catalog fields
      description: Does not change quantityOnHand.
      operationId: updateMedication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                unit: { type: string }
                reorderLevel: { type: integer, minimum: 0 }
                reorderQuantity: { type: integer, minimum: 0 }
                status: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        200:
          description: Updated
        404:
          $ref: "#/components/responses/NotFound"
  /api/medications/{id}/adjust:
    post:
      tags: [Medications]
      summary: Adjust medication stock
      description: "Roles: ADMIN, PHARMACIST. Reason required. Result qty cannot be negative."
      operationId: adjustMedicationStock
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [delta, reason]
              properties:
                delta: { type: integer }
                reason: { type: string, minLength: 3 }
      responses:
        200:
          description: Adjusted
        400:
          description: Invalid delta / negative stock / inactive
        409:
          description: Concurrent update
  /api/medications/{id}/movements:
    get:
      tags: [Medications]
      summary: List stock movements
      operationId: listMedicationMovements
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        200:
          description: Success
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Send header: Authorization: Bearer <token> from login response data.token"
  responses:
    Unauthorized:
      description: "Missing, invalid, or expired Bearer token"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Authentication token is required
    Forbidden:
      description: Authenticated but role not allowed (or patient ownership failed)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: You do not have permission to access this resource
    ValidationError:
      description: Request body failed Zod validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Validation failed
            errors:
              - field: email
                message: Invalid email
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          description: Present on Zod validation failures (HTTP 400)
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...